pyproject.toml
[project]
name = "heurist"
version = "0.1.1"
description = "ETL data pipeline tool for Heurist database."
keywords = [ "duckdb", "etl" ]
dynamic = [ "readme" ]
license = { file = "LICENSE" }
authors = [
    {name = "Kelly Christensen", email = "kelly.christensen@chartes.psl.eu"}
]
requires-python = ">=3.10"
dependencies = [
    "click>=8.2.0",
    "duckdb>=1.2.2",
    "lxml>=5.4.0",
    "pandas>=2.2.3",
    "polars>=1.29.0",
    "pyarrow>=20.0.0",
    "pydantic>=2.11.4",
    "pydantic-xml>=2.16.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "rich>=14.0.0",
    "tenacity>=9.1.2",
]

[project.urls]
Homepage = "https://lostma-erc.github.io/heurist-etl-pipeline/"
Documentation = "https://lostma-erc.github.io/heurist-etl-pipeline/"
Repository = "https://github.com/LostMa-ERC/heurist-etl-pipeline"
Issues = "https://github.com/LostMa-ERC/heurist-etl-pipeline/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
    packages = ["src/heurist", "src/mock_data"]


[tool.poetry.group.dev.dependencies]


[dependency-groups]
dev = [
    "coverage>=7.8.0",
    "genbadge[coverage]>=1.1.2",
    "isort>=6.0.1",
    "mkdocs>=1.6.1",
    "mkdocs-gen-files>=0.5.0",
    "mkdocs-literate-nav>=0.6.2",
    "mkdocs-material[imaging]>=9.6.14",
    "mkdocstrings-python>=1.16.10",
    "pymdown-extensions>=10.15",
    "pytest>=8.3.5",
    "pre-commit>=4.2.0",
    "ruff>=0.11.10",
    "uv>=0.7.5",
]

[project.scripts]
heurist = "heurist.cli.__main__:cli"

[tool.isort]
profile = "black"
line_length = 79
src_paths = ["heurist", "tests"]

[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/legacy/"
publish-url = "https://test.pypi.org/legacy/"

[tool.pytest.ini_options]
testpaths = ["tests", "src/heurist"]
addopts = '--doctest-modules --doctest-glob="*.md"'
doctest_optionflags = [
    "ELLIPSIS",
    "NORMALIZE_WHITESPACE",
    "IGNORE_EXCEPTION_DETAIL"
]
